Remove deprecated _TDBWriter class
